# Vectorize SpanHelpers<T>.IndexOf