--- a/src/libraries/System.Private.CoreLib/src/System/SpanHelpers.T.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/SpanHelpers.T.cs
@@ -226,22 +226,6 @@ namespace System
         {
             Debug.Assert(length >= 0);
 
-            if (typeof(T).IsValueType && RuntimeHelpers.IsBitwiseEquatable<T>())
-            {
-                // bool and char will already have been checked before, just do checks for types
-                // that are equal to sizeof(int) or sizeof(long)
-                if (Unsafe.SizeOf<T>() == sizeof(int))
-                {
-                    int result = IndexOfValueType(ref Unsafe.As<T, int>(ref searchSpace), Unsafe.As<T, int>(ref value), length);
-                    return result != -1;
-                }
-                else if (Unsafe.SizeOf<T>() == sizeof(long))
-                {
-                    int result = IndexOfValueType(ref Unsafe.As<T, long>(ref searchSpace), Unsafe.As<T, long>(ref value), length);
-                    return result != -1;
-                }
-            }
-
             nint index = 0; // Use nint for arithmetic to avoid unnecessary 64->32->64 truncations
 
             if (default(T) != null || (object)value != null)