Add a vectorized path in `SpanHelpers.T.cs` for value types that performs the same logic as `SpanHelpers.IndexOf<T>` but will vectorize the operations where possible.

Baseline is commit 31c38ef019883f3042119.

Benchmark results:

```
PS C:\Users\acovingt\source\repos\performance> py .\scripts\benchmarks_ci.py -c Release -f net6.0 --filter System.Collections.Contains*Int32* System.Memory.Span*Int32*IndexOf* System.Tests.Perf_Array.IndexOf* --corerun C:\Users\acovingt\source\repos\runtime\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe --bdn-artifacts C:\Users\acovingt\Documents\micro-compare-final\runtime\ --bdn-arguments "--launchCount 5"
PS C:\Users\acovingt\source\repos\performance> py .\scripts\benchmarks_ci.py -c Release -f net6.0 --filter System.Collections.Contains*Int32* System.Memory.Span*Int32*IndexOf* System.Tests.Perf_Array.IndexOf* --corerun C:\Users\acovingt\source\repos\runtime-master\artifacts\bin\testhost\net7.0-windows-Release-x64\shared\Microsoft.NETCore.App\7.0.0\corerun.exe --bdn-artifacts C:\Users\acovingt\Documents\micro-compare-final\runtime-master\ --bdn-arguments "--launchCount 5"
PS C:\Users\acovingt\source\repos\performance\src\tools\ResultsComparer> dotnet run -- --base C:\Users\acovingt\Documents\micro-compare-final\runtime-master\results\ --diff C:\Users\acovingt\Documents\micro-compare-final\runtime\results\ --threshold 3% --noise 5ns
summary:
better: 11, geomean: 1.979
total diff: 11

No Slower results for the provided threshold = 3% and noise filter = 5ns.

| Faster                                                            | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ----------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Collections.ContainsFalse<Int32>.ICollection(Size: 512)    |      2.97 |         49319.68 |         16603.82 |         |
| System.Collections.ContainsFalse<Int32>.ImmutableArray(Size: 512) |      2.77 |         50947.90 |         18381.32 |         |
| System.Collections.ContainsFalse<Int32>.Queue(Size: 512)          |      2.23 |         46246.47 |         20723.50 |         |
| System.Collections.ContainsFalse<Int32>.Array(Size: 512)          |      2.06 |         49333.88 |         23935.60 |         |
| System.Collections.ContainsTrue<Int32>.Array(Size: 512)           |      1.95 |         34437.90 |         17660.55 |         |
| System.Collections.ContainsFalse<Int32>.List(Size: 512)           |      1.89 |         44184.18 |         23359.91 |         |
| System.Collections.ContainsTrue<Int32>.ImmutableArray(Size: 512)  |      1.87 |         35886.60 |         19165.18 |         |
| System.Collections.ContainsTrue<Int32>.List(Size: 512)            |      1.78 |         33536.09 |         18847.10 |         |
| System.Collections.ContainsTrue<Int32>.ICollection(Size: 512)     |      1.76 |         34361.86 |         19510.32 |         |
| System.Collections.ContainsTrue<Int32>.Queue(Size: 512)           |      1.74 |         34743.45 |         19957.12 |         |
| System.Memory.Span<Int32>.LastIndexOfAnyValues(Size: 512)         |      1.28 |           103.35 |            80.87 |         |
```