@alexcovington 2021-11-19
Sorry for the delay, @adamsitnik.

> The benchmark numbers look great. Is there any chance you could run them for smaller collections as well?

Absolutely, I've added the following size parameters to the tests you suggested:

```csharp
[Params(7, 25, Utils.DefaultCollectionSize)]
```

Here's the comparison:

```
summary:
better: 26, geomean: 1.814
worse: 6, geomean: 1.129
total diff: 32

| Slower                                                        | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Collections.ContainsTrue<Int32>.Span(Size: 7)          |      1.31 |            22.65 |            29.67 | bimodal |
| System.Collections.ContainsTrue<Int32>.List(Size: 7)          |      1.16 |            39.03 |            45.28 | several?|
| System.Collections.ContainsFalse<Int32>.Array(Size: 7)        |      1.12 |            50.77 |            56.82 |         |
| System.Collections.ContainsFalse<Int32>.Queue(Size: 7)        |      1.09 |            81.53 |            88.55 |         |
| System.Collections.ContainsTrue<Int32>.ImmutableList(Size: 7) |      1.06 |           110.52 |           117.53 |         |
| System.Collections.ContainsFalse<Int32>.HashSet(Size: 25)     |      1.06 |           106.13 |           112.16 | several?|

| Faster                                                               | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| -------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| -------- |
| System.Collections.ContainsFalse<Int32>.Span(Size: 512)              |      3.65 |         50605.50 |         13847.76 |         |
| System.Collections.ContainsFalse<Int32>.List(Size: 512)              |      3.39 |         48622.78 |         14347.67 | several?|
| System.Collections.ContainsFalse<Int32>.Array(Size: 512)             |      3.30 |         49393.34 |         14984.60 | several?|
| System.Collections.ContainsFalse<Int32>.ImmutableArray(Size: 512)    |      3.10 |         51001.78 |         16456.27 |         |
| System.Memory.Span<Int32>.IndexOfValue(Size: 512)                    |      3.05 |            46.63 |            15.31 |         |
| System.Collections.ContainsFalse<Int32>.Queue(Size: 512)             |      3.01 |         50827.41 |         16865.51 | several?|
| System.Collections.ContainsFalse<Int32>.ICollection(Size: 512)       |      3.00 |         44683.99 |         14892.82 |         |
| System.Collections.ContainsTrue<Int32>.Span(Size: 512)               |      2.22 |         32125.17 |         14463.48 |         |
| System.Collections.ContainsTrue<Int32>.List(Size: 512)               |      2.22 |         33579.71 |         15130.16 |         |
| System.Collections.ContainsTrue<Int32>.Queue(Size: 512)              |      2.18 |         34733.56 |         15936.95 |         |
| System.Collections.ContainsTrue<Int32>.ICollection(Size: 512)        |      2.17 |         34269.35 |         15763.35 |         |
| System.Collections.ContainsTrue<Int32>.Array(Size: 512)              |      2.17 |         34494.07 |         15868.69 |         |
| System.Collections.ContainsTrue<Int32>.ImmutableArray(Size: 512)     |      2.08 |         35624.17 |         17141.36 |         |
| System.Collections.ContainsFalse<Int32>.Span(Size: 25)               |      1.62 |           150.01 |            92.59 | several?|
| System.Collections.ContainsTrue<Int32>.List(Size: 25)                |      1.29 |           198.87 |           154.60 |         |
| System.Collections.ContainsFalse<Int32>.List(Size: 25)               |      1.28 |           188.51 |           146.71 |         |
| System.Collections.ContainsTrue<Int32>.Queue(Size: 25)               |      1.26 |           257.97 |           205.55 |         |
| System.Collections.ContainsTrue<Int32>.Array(Size: 25)               |      1.25 |           246.52 |           196.65 |         |
| System.Collections.ContainsFalse<Int32>.ImmutableArray(Size: 25)     |      1.24 |           308.03 |           248.96 |         |
| System.Collections.ContainsTrue<Int32>.Span(Size: 25)                |      1.23 |           135.52 |           110.20 |         |
| System.Collections.ContainsFalse<Int32>.Array(Size: 25)              |      1.22 |           228.54 |           187.44 |         |
| System.Collections.ContainsFalse<Int32>.ICollection(Size: 25)        |      1.17 |           211.66 |           180.44 | several?|
| System.Collections.ContainsTrue<Int32>.ICollection(Size: 25)         |      1.17 |           234.60 |           200.38 |         |
| System.Collections.ContainsTrue<Int32>.ImmutableArray(Size: 25)      |      1.17 |           304.97 |           261.20 |         |
| System.Collections.ContainsTrue<Int32>.ImmutableSortedSet(Size: 512) |      1.07 |         15123.96 |         14125.02 | bimodal |
| System.Collections.ContainsFalse<Int32>.Queue(Size: 25)              |      1.07 |           341.48 |           319.70 |         |
```

So there is a slight regression for smaller buffers that can't be vectorized. It looks like the extra few nanoseconds are due to the extra conditional check that needs to be performed. PerfView shows more samples being collected around one of the checks to see if the type of the buffer is a value type:

<details>

<summary>PerfView screenshots</summary>

![vectorize-array-indexof-baseline-perfview](https://user-images.githubusercontent.com/68252706/142659716-cf06a6a3-72fe-4c85-8b07-c5642fd9bf21.png)

![vectorize-array-indexof-diff-perfview](https://user-images.githubusercontent.com/68252706/142659733-23a1395e-424e-4fc6-8eb2-306c5ab04c0d.png)

</details>

Once the buffer is large enough to be vectorized, the extra cost of the conditional is made up for by the perf increase of the vectorized path.

Please let me know if I can add any additional information or run any other scenarios.