@adamsitnik 2021-11-22
@alexcovington thank you for providing the benchmark numbers. LGTM, :shipit: