@danmoseley 2021-11-22
Nice improvements thanks @alexcovington ! 

Do you plan to do more vectorization type work?