@alexcovington 2021-11-29
Sorry for the delayed response, @danmoseley.

> Do you plan to do more vectorization type work?

Yes, we're always interested in potential vector/SIMD optimization opportunities. Let me know if we can assist with similar work!